requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<3.16
python-dotenv>=1.0,<2.0
bcrypt>=4.0,<5.0
PyJWT>=2.8,<3.0
django-cors-headers>=4.3,<5.0
email-validator>=2.1,<3.0

mongoengine==0.27.0
pymongo==4.8.0
google-auth>=2.34,<3.0
requests>=2.32,<3.0
